// apps/web/lib/api/generated/models/projectResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * llame API
 * llame auth and domain API
 * OpenAPI spec version: 0.1
 */

export interface ProjectResponse {
  id: string;
  ownerUserId: string;
  name: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  archivedAt: string | null;
}